+++ node/resolvers/search/index.ts
@@ -91,6 +91,11 @@ const inputToSearchCrossSelling = {
   [CrossSellingInput.suggestions]: SearchCrossSellingTypes.suggestions,
 }
 
+const getTradePolicyFromSelectedFacets = (selectedFacets: SelectedFacet[]): string | null => {
+  const tradePolicy = selectedFacets.filter(selectedFacet => selectedFacet.key === "trade-policy")
+  return tradePolicy.length > 0 ? tradePolicy[0].value : null
+}
+
 /**
  * There is an URL pattern in VTEX where the number of mapSegments doesn't match the number of querySegments. This function deals with these cases.
  * Since this should not be a search concern, this function will be removed soon.
@@ -557,8 +562,11 @@ export const queries = {
       simulationBehavior,
       hideUnavailableItems,
     } = args
+
     const sellers = await getSellers(vbase, checkout, segment?.channel, segment?.regionId)
     const [count, page] = getProductsCountAndPage(from, to)
+
+    const tradePolicy = getTradePolicyFromSelectedFacets(args.selectedFacets)
 
     const biggyArgs = {
       page,
@@ -579,7 +587,7 @@ export const queries = {
     const productResolver = args.productOriginVtex
       ? productsCatalog
       : productsBiggy
-    const convertedProducts = await productResolver({ ctx, simulationBehavior, searchResult: result })
+    const convertedProducts = await productResolver({ ctx, simulationBehavior, searchResult: result, tradePolicy})
 
     // Add prefix to the cacheId to avoid conflicts. Repeated cacheIds in the same page are causing strange behavior.
     convertedProducts.forEach(product => product.cacheId = `sae-productSearch-${product.cacheId || product.linkText}`)