+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use trade policy from selected facets in `productSearch` query. 
+
 ## [1.28.1] - 2020-12-03
 
 ### Added